+++ src/app/components/aside/aside.component.ts
@@ -5,7 +5,7 @@ import { AccountApiService } from 'src/app/services/api-services/accountApiServi
 import { AccountService } from 'src/app/services/accountService';
 import { ModalService } from 'src/app/services/modalService';
 
-import { Accounts } from '../../pages/accounts/components/cuenta.model';
+import { Accounts } from '../../../models/cuenta.model';
 
 import { NewAccountModal } from '../modals/new-account/new-account';
